refactor of component
