+++ app/react/V2/Routes/Settings/ActivityLog/index.ts
@@ -1,0 +1,2 @@
+export { ActivityLog } from './ActivityLog';
+export { activityLogLoader } from './ActivityLogLoader';
